+++ .travis.yml
@@ -2,8 +2,8 @@ sudo: false
 language: java
 jdk:
     - oraclejdk8
-install: ./gradlew -q assemble
-script: ./gradlew -q test
+install: ./gradlew assemble
+script: ./gradlew test
 cache:
     directories:
     - $HOME/.gradle/caches/
