--- src/test/java/com/losd/reqbot/repository/BucketRedisRepoTest.java
@@ -1,68 +1,0 @@
-package com.losd.reqbot.repository;
-
-import com.losd.reqbot.config.JedisConfiguration;
-import com.losd.reqbot.config.RepoConfiguration;
-import org.junit.Before;
-import org.junit.Ignore;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import redis.clients.jedis.Jedis;
-
-import java.util.Set;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.hasItems;
-import static org.hamcrest.Matchers.hasSize;
-
-/**
- * The MIT License (MIT)
- * <p>
- * Copyright (c) 2015 Andrew Braithwaite
- * <p>
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * <p>
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- * <p>
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = {RedisConfiguration.class, RepoConfiguration.class, JedisConfiguration.class})
-public class BucketRedisRepoTest {
-    @Autowired
-    BucketRepo repo;
-
-    @Autowired
-    Jedis jedis;
-
-    @Before
-    public void setup() {
-        jedis.flushDB();
-    }
-
-    @Test
-    public void it_can_get_a_set_containing_all_of_the_buckets() {
-        jedis.lpush(BucketRedisRepo.getBucketKey("a"), "element");
-        jedis.lpush(BucketRedisRepo.getBucketKey("b"), "element");
-        jedis.lpush(BucketRedisRepo.getBucketKey("c"), "element");
-        jedis.lpush(BucketRedisRepo.getBucketKey("d"), "element");
-
-        Set<String> buckets = repo.getBuckets();
-        assertThat(buckets, hasSize(4));
-        assertThat(buckets, hasItems("a", "b", "c", "d"));
-    }
-}
